consolidated database access to dbConfig js file to draw all clients from one pool, also deleted commented out code
